+++ fe/src/service/axios/mood/mood.ts
@@ -1,5 +1,5 @@
-import { publicApi } from 'api/fetcher';
-import { END_POINT } from 'constants/endpoint';
+import { publicApi } from 'service/axios/fetcher';
+import { END_POINT } from 'service/constants/endpoint';
 
 export const getTasteMoods = async () => {
   const { data } = await publicApi.get(END_POINT.mood);
