+++ fe/src/service/axios/profile/profile.ts
@@ -1,5 +1,5 @@
-import { privateApi } from 'api/fetcher';
-import { END_POINT } from 'constants/endpoint';
+import { privateApi } from 'service/axios/fetcher';
+import { END_POINT } from 'service/constants/endpoint';
 
 export const getProfile = async (memberId?: string) => {
   const { data } = await privateApi.get(END_POINT.member(memberId));
